Compress BondData
Times, status and claimed amount into 1 slot.

+++ LUSDChickenBonds/src/Interfaces/IChickenBondManager.sol
@@ -30,7 +30,7 @@ interface IChickenBondManager {
 
     // getters
     function calcRedemptionFeePercentage(uint256 _fractionOfBLUSDToRedeem) external view returns (uint256);
-    function getBondData(uint256 _bondID) external view returns (uint256 lusdAmount, uint256 claimedBLUSD, uint256 startTime, uint256 endTime, uint8 status);
+    function getBondData(uint256 _bondID) external view returns (uint256 lusdAmount, uint64 claimedBLUSD, uint64 startTime, uint64 endTime, uint8 status);
     function getLUSDToAcquire(uint256 _bondID) external view returns (uint256);
     function calcAccruedBLUSD(uint256 _bondID) external view returns (uint256);
     function calcBondBLUSDCap(uint256 _bondID) external view returns (uint256);
